+++ docs/CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file. The format 
 - Optional usage tracking (github issue #200)
 
 ### Changed
-- The template bidsmaps are now stored in `[home]/.bidscoin/templates`
+- The template bidsmaps are now stored in `[home]/.bidscoin/[version]/templates`
 
 ## [4.1.1] - 2023-09-14